bug fixes, refactoring, move extended view logic to new server module
