+++ fiftyone/server/extended_view.py
@@ -1,0 +1,260 @@
+"""
+FiftyOne extended view.
+
+| Copyright 2017-2021, Voxel51, Inc.
+| `voxel51.com <https://voxel51.com/>`_
+|
+"""
+from fiftyone.core.expressions import ViewField as F, VALUE
+import fiftyone.core.fields as fof
+import fiftyone.core.labels as fol
+import fiftyone.core.media as fom
+import fiftyone.core.stages as fosg
+import fiftyone.core.state as fos
+
+
+_BOOL_FILTER = "bool"
+_NUMERIC_FILTER = "numeric"
+_STR_FILTER = "str"
+
+_LABEL_TAGS = "_label_tags"
+
+
+def get_extended_view(view, filters, count_labels_tags=False):
+    """Create an extended view with the provided filters.
+
+    Args:
+        view: a :class:`fiftyone.core.collections.SampleCollection`
+        filters: a `dict` of App defined filters
+        count_labels_tags (False): whether to set the hideen `_label_tags` field
+            with counts of tags with respect to all label fields
+    """
+    cleanup_pipeline = None
+
+    if filters is not None and len(filters):
+        filters = filters.copy()
+
+        label_tags = None
+        if "tags" in filters:
+            tags = filters.pop("tags")
+            if "label" in tags:
+                label_tags = tags["label"]
+
+            if "sample" in tags:
+                view = view.match_tags(tags=tags["sample"])
+
+        stages, cleanup_pipeline = _make_filter_stages(
+            view, filters, label_tags=label_tags, hide_result=count_labels_tags
+        )
+
+        for stage in stages:
+            view = view.add_stage(stage)
+
+    if count_labels_tags:
+        view = _add_labels_tags_counts(view, filters)
+        if cleanup_pipeline:
+            view = view.mongo(cleanup_pipeline)
+
+    return view
+
+
+def _add_labels_tags_counts(view, filters):
+    fields = fos.DatasetStatistics.labels(view)
+    view = view.set_field(_LABEL_TAGS, [])
+    for path, field in fields:
+        if not issubclass(
+            field.document_type, (fol._HasID, fol._HasLabelList)
+        ):
+            continue
+
+        path = _get_filtered_path(view, path, filters)
+        if issubclass(field.document_type, fol._HasLabelList):
+            if path.startswith(view._FRAMES_PREFIX):
+                add_tags = _add_frame_labels_tags
+            else:
+                add_tags = _add_labels_tags
+        else:
+            if path.startswith(view._FRAMES_PREFIX):
+                add_tags = _add_frame_label_tags
+            else:
+                add_tags = _add_label_tags
+
+        view = add_tags(path, field, view)
+
+    view = _count_list_items(_LABEL_TAGS, view)
+
+    return view
+
+
+def _make_scalar_expression(f, args):
+    expr = None
+    cls = args["_CLS"]
+    if cls == _BOOL_FILTER:
+        true, false = args["true"], args["false"]
+        if true and false:
+            expr = f.is_in([True, False])
+
+        if not true and false:
+            expr = f == False
+
+        if true and not false:
+            expr = f == True
+
+        if not true and not false:
+            expr = (f != True) & (f != False)
+
+    elif cls == _NUMERIC_FILTER:
+        mn, mx = args["range"]
+        expr = (f >= mn) & (f <= mx)
+    elif cls == _STR_FILTER:
+        values = args["values"]
+        if not values:
+            return None
+
+        none = any(map(lambda v: v is None, values))
+        values = filter(lambda v: v is not None, values)
+        expr = f.is_in(values)
+        exclude = args["exclude"]
+
+        if exclude:
+            # pylint: disable=invalid-unary-operand-type
+            expr = ~expr
+
+        if none:
+            if exclude:
+                expr &= f.exists()
+            else:
+                expr |= ~(f.exists())
+
+        return expr
+
+    none = args["none"]
+    if not none:
+        if expr is not None:
+            expr &= f.exists()
+        else:
+            expr = f.exists()
+    elif expr is not None:
+        expr |= ~(f.exists())
+
+    return expr
+
+
+def _make_filter_stages(view, filters, label_tags=None, hide_result=False):
+    field_schema = view.get_field_schema()
+    if view.media_type == fom.VIDEO:
+        frame_field_schema = view.get_frame_field_schema()
+    else:
+        frame_field_schema = None
+
+    tag_expr = (F("tags") != None).if_else(
+        F("tags").contains(label_tags), False
+    )
+
+    stages = []
+    cleanup = []
+    for path, args in filters.items():
+        keys = path.split(".")
+        frames = path.startswith(view._FRAMES_PREFIX)
+        if frames:
+            schema = frame_field_schema
+            field = schema[keys[1]]
+            path = ".".join(keys[:2])
+        else:
+            schema = field_schema
+            path = keys[0]
+            field = schema[path]
+
+        if isinstance(field, fof.EmbeddedDocumentField):
+            expr = _make_scalar_expression(F(keys[-1]), args)
+            if label_tags:
+                if expr is not None:
+                    expr &= tag_expr
+                else:
+                    expr = tag_expr
+
+            if expr is not None:
+                if hide_result:
+                    new_field = "__%s" % field
+                    if frames:
+                        new_field = "%s%s" % (view._FRAMES_PREFIX, new_field,)
+                else:
+                    new_field = None
+                stages.append(
+                    fosg.FilterLabels(path, expr, _new_field=new_field)
+                )
+                cleanup.append({"$unset": path})
+        else:
+            expr = _make_scalar_expression(F(path), args)
+            if expr is not None:
+                stages.append(fosg.Match(expr))
+
+    if label_tags is not None:
+        for path, _ in fos.DatasetStatistics.labels(view):
+            if path not in filters:
+                stages.append(fosg.FilterLabels(path, tag_expr))
+
+    return stages, cleanup
+
+
+def _get_filtered_path(view, path, filters):
+    if path not in filters:
+        return path
+
+    if path.startswith(view._FRAMES_PREFIX):
+        return "%s__%s" % (view._FRAMES_PREFIX, path.split(".")[1])
+
+    return "__%s" % path
+
+
+def _add_frame_labels_tags(path, field, view):
+    frames, path = path.split(".")
+    items = "%s.%s" % (path, field.document_type._LABEL_LIST_FIELD)
+    view = view.set_field(
+        _LABEL_TAGS,
+        F(_LABEL_TAGS).extend(
+            F(frames).reduce(
+                VALUE.extend(F(items).reduce(VALUE.extend(F("tags")), [])), []
+            )
+        ),
+    )
+    return view
+
+
+def _add_frame_label_tags(path, field, view):
+    frames, path = path.split(".")
+    tags = "%s.tags" % path
+    view = view.set_field(
+        _LABEL_TAGS,
+        F(_LABEL_TAGS).extend(F(frames).reduce(VALUE.extend(F(tags)), [])),
+    )
+    return view
+
+
+def _add_labels_tags(path, field, view):
+    items = "%s.%s" % (path, field.document_type._LABEL_LIST_FIELD)
+    view = view.set_field(
+        _LABEL_TAGS,
+        F(_LABEL_TAGS).extend(F(items).reduce(VALUE.extend(F("tags")), [])),
+    )
+    return view
+
+
+def _add_label_tags(path, field, view):
+    return view.set_field(
+        _LABEL_TAGS, F(_LABEL_TAGS).extend(F("%s.tags" % path))
+    )
+
+
+def _count_list_items(path, view):
+    function = (
+        "function(items) {"
+        "let counts = {};"
+        "items.forEach((i) => {"
+        "counts[i] = 1 + (counts[i] || 0);"
+        "});"
+        "return counts;"
+        "}"
+    )
+
+    return view.set_field(path, F(path)._function(function))
